v1/backend/requirements.txt
fastapi
uvicorn
playwright
langchain
langchain-openai
python-dotenv
sqlmodel
boto3
google-cloud-texttospeech
openai
pywinauto
uiautomation
